Move async test to py3 only
