+++ python/ray/tests/py3_test.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import asyncio
 import pytest
 
 import ray
@@ -94,3 +95,32 @@ def test_args_intertwined(ray_start_regular):
     local_method = local_actor.cls_args_intertwined
     test_function(local_method, actor_method)
     ray.get(remote_test_function.remote(local_method, actor_method))
+
+def test_asyncio_actor(ray_start_regular):
+    @ray.remote
+    class AsyncBatcher(object):
+        def __init__(self):
+            self.batch = []
+            # The event currently need to be created from the same thread.
+            # We currently run async coroutines from a different thread.
+            self.event = None
+
+        async def add(self, x):
+            if self.event is None:
+                self.event = asyncio.Event()
+            self.batch.append(x)
+            if len(self.batch) >= 3:
+                self.event.set()
+            else:
+                await self.event.wait()
+            return sorted(self.batch)
+
+    a = AsyncBatcher.options(is_direct_call=True, is_asyncio=True).remote()
+    x1 = a.add.remote(1)
+    x2 = a.add.remote(2)
+    x3 = a.add.remote(3)
+    r1 = ray.get(x1)
+    r2 = ray.get(x2)
+    r3 = ray.get(x3)
+    assert r1 == [1, 2, 3]
+    assert r1 == r2 == r3
